Add additional build-deps
If the system that built libnbcompat is different from the one building
mtree-netbsd in terms of whether it has libncurses-dev or libbsd-dev
installed, compilation errors will result due to differing opinions
on whether, for instance, termcap.h is present.

I could either build-conflict or build-depend on these packages.
Opting to build-dep as the most sensible approach.